Adding latest stuff.
Unfortunately, this is way too big. I lost track of some stuff during vacation, trying to get back
track here. What I can say is that the streaming stuff is getting there, but certainly not usable,
and the unit tests are not all there either. But the base algorithm is sound.