Changes to VarInt operations and some refactoring for clarity.
VarInt operations now done in their own class rather than in (de)compressor.
Renamed the parameters in the FindMatchLength methods to be more descriptive.
Removed fast path from EmitLiteral as I am not sure it helps. Want to investigate if it is worthwhile at a later time.
EmitLiteralTagBytes is now more direct rather than a loop that builds the right thing opaquely.